[RDF] Move Snapshot-related functions / types to their own files.
Split all Snapshot-related things from the rest of the RDF action
helpers.
Used include-what-you-use to minimise the amount of dead includes in the
affected files.
