+++ tree/dataframe/src/RDFInterfaceUtils.cxx
@@ -13,7 +13,6 @@
 #include <ROOT/RDF/InterfaceUtils.hxx>
 #include <ROOT/RDF/RColumnRegister.hxx>
 #include <ROOT/RDF/RDisplay.hxx>
-#include <ROOT/RDF/RInterface.hxx>
 #include <ROOT/RDF/RJittedDefine.hxx>
 #include <ROOT/RDF/RJittedFilter.hxx>
 #include <ROOT/RDF/RJittedVariation.hxx>
@@ -56,26 +55,15 @@
 #include <sstream>
 #include <stdexcept>
 #include <string>
-#include <type_traits> // for remove_reference<>::type
 #include <typeinfo>
 #include <unordered_map>
 #include <unordered_set>
 #include <utility> // for pair
 #include <vector>
 
-namespace ROOT {
-namespace Detail {
-namespace RDF {
+namespace ROOT::Detail::RDF {
 class RDefineBase;
-} // namespace RDF
-namespace Internal {
-namespace RDF {
-class RJittedAction;
 }
-} // namespace Internal
-} // namespace Detail
-
-} // namespace ROOT
 
 namespace {
 using ROOT::Internal::RDF::IsStrInVec;
